Handle track and check command

+++ db/00-init.sql
@@ -5,6 +5,6 @@ CREATE TABLE TrackingNumbers (
 	id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
 	currentState integer NOT NULL DEFAULT 0,
 	totalState integer NOT NULL DEFAULT 0,
-	chatId integer NOT NULL,
+	chatId bigint NOT NULL,
 	trackingNumber varchar(30) NOT NULL DEFAULT ''
 )
